Moving implementation specific classes to Implementation folder
